datatrails: refactor: move related code to trails/MetricSelect subdir (#88262)
refactor: move related code to trails/MetricSelect subdir
