+++ public/app/features/trails/ActionTabs/RelatedMetricsScene.tsx
@@ -1,4 +1,4 @@
-import { MetricSelectScene } from '../MetricSelectScene';
+import { MetricSelectScene } from '../MetricSelect/MetricSelectScene';
 
 export function buildRelatedMetricsScene() {
   return new MetricSelectScene({});
