+++ public/app/features/trails/MetricSelect/hideEmptyPreviews.ts
