+++ public/app/features/trails/MetricSelect/previewPanel.test.ts
